modified: ItemService.java - add @Slf4j (createItem, read) - Return Type(createItem, updateItem): SalesItem(dto) -> void(cuz Controller) - updateItem -> Implement Check Password - deleteItem -> Implement Check Password

+++ src/main/java/com/be05/market/service/ItemService.java
@@ -4,6 +4,7 @@ import com.be05.market.dto.SalesItem;
 import com.be05.market.entity.ItemEntity;
 import com.be05.market.repository.ItemRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -14,12 +15,14 @@ import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Optional;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class ItemService {
     private final ItemRepository itemRepository;
 
-    public SalesItem createItem(SalesItem items) {
+    // CREATE
+    public void createItem(SalesItem items) {
         ItemEntity newItems = new ItemEntity();
         newItems.setTitle(items.getTitle());
         newItems.setDescription(items.getDescription());
@@ -27,37 +30,55 @@ public class ItemService {
         newItems.setWriter(items.getWriter());
         newItems.setPassword(items.getPassword());
         newItems.setStatus("판매중");
-        return SalesItem.fromEntity(itemRepository.save(newItems));
+        SalesItem.fromEntity(itemRepository.save(newItems));
+        log.info(String.valueOf(newItems));
     }
 
+    // FindAll(Pages)
     public Page<SalesItem> readItemsPaged(Integer page, Integer limit) {
-        Pageable pageable = PageRequest.of(page, limit, Sort.by("id").descending());
+        Pageable pageable =
+                PageRequest.of(page, limit, Sort.by("id").descending());
         Page<ItemEntity> itemEntities = itemRepository.findAll(pageable);
         return itemEntities.map(SalesItem::fromEntity);
     }
 
+    // FindById
     public SalesItem read(Long id) {
         Optional<ItemEntity> optionalItem = itemRepository.findById(id);
+        log.info(String.valueOf(optionalItem));
         if (optionalItem.isPresent()) return SalesItem.fromEntity(optionalItem.get());
         else throw new ResponseStatusException(HttpStatus.NOT_IMPLEMENTED);
     }
 
-    public SalesItem updateItem(Long id, SalesItem item) {
+    // Update
+    public void updateItem(Long id, SalesItem item) {
         Optional<ItemEntity> optionalItem = itemRepository.findById(id);
         if (optionalItem.isPresent()) {
             ItemEntity itemEntity = optionalItem.get();
-            itemEntity.setTitle(item.getTitle());
-            itemEntity.setDescription(item.getDescription());
-            itemEntity.setMinPriceWanted(item.getMinPriceWanted());
-            itemRepository.save(itemEntity);
-            return SalesItem.fromEntity(itemEntity);
-        }
-        else throw new ResponseStatusException(HttpStatus.NOT_IMPLEMENTED);
+
+            // Check Password
+            if (itemEntity.getPassword().equals(item.getPassword())) {
+                itemEntity.setTitle(item.getTitle());
+                itemEntity.setDescription(item.getDescription());
+                itemEntity.setMinPriceWanted(item.getMinPriceWanted());
+                itemRepository.save(itemEntity);
+            } else throw new ResponseStatusException(HttpStatus.NOT_IMPLEMENTED);
+        } else throw new ResponseStatusException(HttpStatus.NOT_FOUND);
     }
 
-    public void deleteItem(Long id) {
-        if (itemRepository.existsById(id))
-            itemRepository.deleteById(id);
-        else throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+    // Delete
+    public void deleteItem(Long id, SalesItem item) {
+        Optional<ItemEntity> optionalItem = itemRepository.findById(id);
+        if (optionalItem.isPresent()) {
+            ItemEntity itemEntity = optionalItem.get();
+
+            if (itemEntity.getPassword().equals(item.getPassword())) {
+                if (itemRepository.existsById(id)) {
+                    itemRepository.deleteById(id);
+                }
+                else throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+            }
+        } else throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+
     }
 }
